// src/types.ts
namespace Types {
  export interface MovieInterface {
    title: string;
    release_year: number;
    locations: string[];
    fun_facts: string;
    production_company: string;
    distributor: string;
    director: string;
    writer: string;
    actors: string[];
  }
}

export default Types;